Revert "Updated floats to doubles to reduce confusion."
This reverts commit 590522eb5bfdbf4c26ad2a312584283ce550e133.

+++ src/ocean_models/include/fvcom/FVCOMChunk.h
@@ -14,16 +14,16 @@ class FVCOMChunk
 public:
 	struct NodeData
 	{
-		double temp;
-		double salt;
-		double dye;
+		float temp;
+		float salt;
+		float dye;
 	};
 
 	struct TriangleData
 	{
-		double u;
-		double v;
-		double w;
+		float u;
+		float v;
+		float w;
 	};
 
 	typedef std::vector<FVCOMChunk::NodeData> NodeVector;
